reapply of alpha fixes

+++ compiler/sc1.c
@@ -5,7 +5,7 @@
  *  Copyright (c) ITB CompuPhase, 1997-2002
  *  This file may be freely used. No warranties of any kind.
  *
- *  Version: $Id: sc1.c,v 1.6 2003/07/17 22:21:57 sparhawksf Exp $
+ *  Version: $Id: sc1.c,v 1.7 2003/07/18 17:49:01 dgp85 Exp $
  */
 #include <assert.h>
 #include <ctype.h>
@@ -935,8 +935,8 @@ static void setconstants(void)
     add_constant("cellmin",SHRT_MIN,sGLOBAL,0);
   #else
     add_constant("cellbits",32,sGLOBAL,0);
-    add_constant("cellmax",LONG_MAX,sGLOBAL,0);
-    add_constant("cellmin",LONG_MIN,sGLOBAL,0);
+    add_constant("cellmax",INT_MAX,sGLOBAL,0);
+    add_constant("cellmin",INT_MIN,sGLOBAL,0);
   #endif
   add_constant("charbits",charbits,sGLOBAL,0);
   add_constant("charmin",0,sGLOBAL,0);
